feat(asset-reminders): add sorting filter on asset reminders page
